requirements.txt
# NanoGUI Requirements
# ===================
# Install with: pip install -r requirements.txt

# Core Dependencies
# ----------------
autogen-agentchat>=0.4.0
autogen-core>=0.4.0
autogen-ext[openai]>=0.4.0  # For OpenAI-compatible clients (GLM)

# Machine Learning
# ----------------
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
peft>=0.8.0
accelerate>=0.25.0
bitsandbytes>=0.42.0  # For 4-bit quantization
qwen-vl-utils>=0.0.1  # For Qwen2-VL image processing

# Computer Vision
# ---------------
pillow>=10.0.0
numpy>=1.24.0

# GUI Automation
# ---------------
pyautogui>=0.9.54
mss>=9.0.1  # Cross-platform screen capture

# Data Handling
# --------------
datasets>=2.16.0  # HuggingFace datasets
pandas>=2.0.0
pyyaml>=6.0.0  # YAML configuration files

# Training & Logging
# ------------------
wandb>=0.16.0  # Weights & Biases (optional, for experiment tracking)
# tensorboard>=2.15.0  # Alternative to wandb

# Evaluation
# -----------
scikit-learn>=1.3.0
matplotlib>=3.8.0

# Development
# ------------
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.3.0
mypy>=1.8.0

# Optional: Fast API for serving
# -------------------------------
# fastapi>=0.110.0
# uvicorn>=0.28.0
# gradio>=4.0.0  # For web UI

# Platform-specific (uncomment if needed)
# ---------------------------------------
# pywin32>=306  # Windows-specific (for pyautogui)
# pyscreenshot>=3.0  # Alternative screen capture
